requirements.txt
Flask==3.1.0
requests==2.32.3
beautifulsoup4==4.12.3
Pillow==11.1.0
